requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Clientes IA
anthropic==0.26.0
openai==1.12.0
google-generativeai==0.3.2

# Procesamiento de documentos
python-docx==1.1.0
reportlab==4.0.9
Pillow==10.2.0

# Utilidades
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Desarrollo (opcional)
pytest==8.0.0
black==24.1.1
flake8==7.0.0
